requirements.txt
# flake8 stuff
flake8==3.7.9
flake8-annotations-complexity==0.0.2
flake8-broken-line==0.1.1
flake8-bugbear==19.8.0
flake8-builtins==1.4.1
flake8-class-attributes-order==0.0.7
flake8-commas==2.0.0
flake8-comprehensions==3.1.4
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-functions==0.0.2
flake8-polyfill==1.0.2
flake8-print==3.1.4
flake8-quotes==2.1.1
flake8-string-format==0.2.3
flake8-todo==0.7
flake8-variables-names==0.0.3
flake8-cognitive-complexity==0.0.2

# mypy stuff
mypy==0.750
mypy-extensions==0.4.3

# pytest stuff
pytest==5.3.1
pytest-cov==2.8.1
pytest-env==0.6.2
pytest-mock==1.13.0
pytest-randomly==3.1.0
pytest-socket==0.3.3
freezegun==0.3.12
factory_boy==2.12.0